refactor: 리팩토링 - 패키지 구조 수정 - 코드 컨벤션에 맞게 포맷팅 - 상수 수정
