+++ src/main/java/carracing/Validation.java
@@ -1,6 +1,5 @@
 package carracing;
 
-
 import java.util.regex.Pattern;
 
 import static java.util.Objects.*;
