+++ src/main/java/carracing/model/MovingStrategy.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.model;
 
 public interface MovingStrategy {
     int move();
