+++ src/main/java/carracing/model/Race.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.model;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -20,7 +20,7 @@ public class Race {
     private List<Car> participatedCarList() {
         List<Car> participatedCarList = new ArrayList<>();
 
-        for(int i = 0; i < numberOfCars; i++) {
+        for (int i = 0; i < numberOfCars; i++) {
             participatedCarList.add(new Car(new RandomStrategy()));
         }
 
@@ -30,7 +30,7 @@ public class Race {
     private List<RaceStatus> doRace(List<Car> carList) {
         List<RaceStatus> raceStatusList = new ArrayList<>();
 
-        for(int i = 0; i < numberOfAttempts; i++) {
+        for (int i = 0; i < numberOfAttempts; i++) {
             boolean isLastAttempt = i == (numberOfAttempts - 1);
             carList.forEach(Car::move);
             raceStatusList.add(raceStatus(carList, isLastAttempt));
