+++ src/test/java/carracing/ValidationTest.java
@@ -1,6 +1,5 @@
 package carracing;
 
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
