+++ src/test/java/carracing/model/RaceStatusTest.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
